# web-backend/openapi.yaml
openapi: 3.0.3
info:
  title: Graph Learning System API
  description: |
    API for the Graph Learning System (abcdeez) - an adaptive cognitive learning platform.
    
    ## Authentication
    Most endpoints require JWT authentication. Obtain a token via `/auth/login`.
    
    ## Rate Limiting
    - Default: 100 requests/minute/IP
    - Authenticated: 1000 requests/minute/user
  version: 1.0.0
  contact:
    email: api@graphlearning.app
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: https://api.graphlearning.app/v1
    description: Production server
  - url: https://staging-api.graphlearning.app/v1
    description: Staging server
  - url: http://localhost:8080/v1
    description: Local development

tags:
  - name: auth
    description: Authentication operations
  - name: sessions
    description: Learning session management
  - name: tasks
    description: Task generation and responses
  - name: analytics
    description: Performance analytics
  - name: export
    description: Data export operations

paths:
  /auth/register:
    post:
      tags: [auth]
      summary: Register new user
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '422':
          $ref: '#/components/responses/ValidationError'

  /auth/login:
    post:
      tags: [auth]
      summary: Login user
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /sessions:
    post:
      tags: [sessions]
      summary: Create learning session
      operationId: createSession
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
      responses:
        '201':
          description: Session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    
    get:
      tags: [sessions]
      summary: List sessions
      operationId: listSessions
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: status
          in: query
          schema:
            type: string
            enum: [active, completed, abandoned]
        - name: domain
          in: query
          schema:
            type: string
        - name: from_date
          in: query
          schema:
            type: string
            format: date-time
        - name: to_date
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Sessions retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionList'

  /sessions/{sessionId}:
    get:
      tags: [sessions]
      summary: Get session details
      operationId: getSession
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/sessionId'
      responses:
        '200':
          description: Session details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionDetail'
        '404':
          $ref: '#/components/responses/NotFoundError'

  /sessions/{sessionId}/tasks:
    post:
      tags: [tasks]
      summary: Get next task
      operationId: getNextTask
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/sessionId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreviousResponse'
      responses:
        '200':
          description: Next task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'

  /sessions/{sessionId}/tasks/{taskId}/response:
    post:
      tags: [tasks]
      summary: Submit task response
      operationId: submitResponse
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/sessionId'
        - $ref: '#/components/parameters/taskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskResponse'
      responses:
        '200':
          description: Response recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseResult'

  /users/{userId}/analytics:
    get:
      tags: [analytics]
      summary: Get user analytics
      operationId: getUserAnalytics
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/userId'
        - name: from_date
          in: query
          schema:
            type: string
            format: date-time
        - name: to_date
          in: query
          schema:
            type: string
            format: date-time
        - name: domain
          in: query
          schema:
            type: string
        - name: granularity
          in: query
          schema:
            type: string
            enum: [day, week, month]
            default: day
      responses:
        '200':
          description: Analytics data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Analytics'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    sessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    taskId:
      name: taskId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    userId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid

  schemas:
    RegisterRequest:
      type: object
      required: [email, password, display_name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        display_name:
          type: string

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
        expires_in:
          type: integer
        user:
          $ref: '#/components/schemas/User'

    User:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        display_name:
          type: string
        created_at:
          type: string
          format: date-time

    CreateSessionRequest:
      type: object
      required: [domain]
      properties:
        domain:
          type: string
          enum: [alphabet, numbers, music, days]
        config:
          $ref: '#/components/schemas/SessionConfig'

    SessionConfig:
      type: object
      properties:
        difficulty:
          type: number
          minimum: 0
          maximum: 1
        hint_probability:
          type: number
          minimum: 0
          maximum: 1
        max_trials:
          type: integer
          minimum: 1

    Session:
      type: object
      properties:
        session_id:
          type: string
          format: uuid
        domain:
          type: string
        config:
          $ref: '#/components/schemas/SessionConfig'
        started_at:
          type: string
          format: date-time
        status:
          type: string
          enum: [active, completed, abandoned]

    SessionDetail:
      allOf:
        - $ref: '#/components/schemas/Session'
        - type: object
          properties:
            ended_at:
              type: string
              format: date-time
            metrics:
              $ref: '#/components/schemas/SessionMetrics'
            trials:
              type: array
              items:
                $ref: '#/components/schemas/Trial'

    SessionMetrics:
      type: object
      properties:
        total_trials:
          type: integer
        correct_trials:
          type: integer
        accuracy:
          type: number
        average_rt:
          type: number
        rt_variance:
          type: number
        learning_curve:
          type: array
          items:
            type: number

    SessionList:
      type: object
      properties:
        sessions:
          type: array
          items:
            $ref: '#/components/schemas/Session'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer

    Task:
      type: object
      properties:
        task_id:
          type: string
          format: uuid
        session_id:
          type: string
          format: uuid
        stimulus:
          type: string
        choices:
          type: array
          items:
            type: string
        task_type:
          type: string
        difficulty:
          type: number
        hint:
          type: string
          nullable: true

    PreviousResponse:
      type: object
      properties:
        task_id:
          type: string
          format: uuid
        response:
          type: string
        response_time:
          type: number
        correct:
          type: boolean

    TaskResponse:
      type: object
      required: [response, response_time]
      properties:
        response:
          type: string
        response_time:
          type: number
        client_timestamp:
          type: string
          format: date-time

    ResponseResult:
      type: object
      properties:
        task_id:
          type: string
          format: uuid
        correct:
          type: boolean
        expected:
          type: string
        feedback:
          type: string
        next_task_available:
          type: boolean

    Trial:
      type: object
      properties:
        trial_id:
          type: string
          format: uuid
        task_id:
          type: string
          format: uuid
        stimulus:
          type: string
        response:
          type: string
        correct:
          type: boolean
        response_time:
          type: number
        timestamp:
          type: string
          format: date-time

    Analytics:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        period:
          type: object
          properties:
            from:
              type: string
              format: date-time
            to:
              type: string
              format: date-time
        summary:
          type: object
          properties:
            total_sessions:
              type: integer
            total_time_minutes:
              type: number
            average_accuracy:
              type: number
            improvement_rate:
              type: number
            domains_practiced:
              type: array
              items:
                type: string
        time_series:
          type: array
          items:
            type: object
        domain_breakdown:
          type: object
          additionalProperties:
            type: object

    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
            timestamp:
              type: string
              format: date-time
            request_id:
              type: string
              format: uuid

  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    
    RateLimitError:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
        X-RateLimit-Reset:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'